+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidauth/__init__.py
@@ -1,0 +1,4 @@
+"""SquidGuard plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidauth import Plugin  # NOQA