+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidcache/__init__.py
@@ -1,0 +1,4 @@
+"""Squid Cache plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidcache import Plugin  # NOQA