+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidguardacl/__init__.py
@@ -1,0 +1,4 @@
+"""SquidGuard ACL plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidguardacl import Plugin  # NOQA