+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidguarddefault/__init__.py
@@ -1,0 +1,4 @@
+"""SquidGuard Default plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidguarddefault import Plugin  # NOQA