+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidguardrewrite/__init__.py
@@ -1,0 +1,4 @@
+"""SquidGuard Rewrite plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidguardrewrite import Plugin  # NOQA