+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidguardtime/__init__.py
@@ -1,0 +1,4 @@
+"""SquidiGuard Time plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidguardtime import Plugin  # NOQA