+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidnac/__init__.py
@@ -1,0 +1,4 @@
+"""Squid NAC plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidnac import Plugin  # NOQA