+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidremote/__init__.py
@@ -1,0 +1,4 @@
+"""Squid Remote plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidremote import Plugin  # NOQA