+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidreversegeneral/__init__.py
@@ -1,0 +1,4 @@
+"""Squid Reverse (gen) plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidreversegeneral import Plugin  # NOQA