+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidreversepeer/__init__.py
@@ -1,0 +1,4 @@
+"""Squid Reverse (Peer) plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidreversepeer import Plugin  # NOQA