+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_squidtraffic/__init__.py
@@ -1,0 +1,4 @@
+"""Squid traffic."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .squidtraffic import Plugin  # NOQA